utils: add support for index files
Add the ability to define "index" files -- that is, files that
get served even when no file is defined.

For example, if you have a subdirectory called "stuff" that contains
a file called "default.html", and you would like to serve "default.html"
when accessing /stuff/ (or /stuff), specify an "index" of "default.html"
when deploying.

Signed-off-by: Sean Cross <sean@xobs.io>